Adding importing functions
Core functions for importing added.
Importing static meshes requires a change of the settings.
There is a bug while changing the settings of the textures: the engine crashes when trying to disable virtual texture streaming and change compression. Each operation can be successfully proceeded separately, but trying to do both results in a crash. Saving dirty packages might help.